requirements.txt
PyMuPDF>=1.24.0
scikit-learn>=1.3.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pytest>=7.0.0
# Auth + DB
sqlalchemy>=2.0.0
alembic>=1.12.0
argon2-cffi>=23.1.0
python-multipart>=0.0.6
# DB drivers (Postgres)
psycopg2-binary>=2.9.0
